приведение количества полок к int

+++ vlasov_ms/Practice/Practice1/main.c
@@ -1,6 +1,5 @@
 ﻿#include <stdio.h>
 #include <locale.h>
-#include <math.h>
 
 void main() 
 {
@@ -60,7 +59,7 @@ void main()
     vBok = 2 * h * d * 0.015;
     vKry = 2 * w * d * 0.015;
     vDve = h * w * 0.01;
-    vPol = (h - 2 * 0.015) / (0.4 + 0.015) * (w - 2 * 0.015) * (d - 0.005) * 0.015;
+    vPol = (int)((h - 2 * 0.015) / (0.4 + 0.015)) * (w - 2 * 0.015) * (d - 0.005) * 0.015;
     m = vZad * pDvp + vBok * pDsp + vKry * pDsp + vDve * pWood + vPol * pDsp;
     printf("\nМАССА ШКАФА: %.2lf кг\n", m);
 }
